Fix linter warnings
