+++ Sources/TokamakDOM/Views/Canvas/ImageOperations.swift
@@ -1,0 +1,28 @@
+// Copyright 2020-2021 Tokamak contributors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+//  Created by Carson Katri on 9/23/21.
+//
+
+import Foundation
+import TokamakCore
+
+extension _Canvas {
+  func resolveImage(
+    _ image: Image,
+    _ environment: EnvironmentValues
+  ) -> GraphicsContext.ResolvedImage {
+    ._resolved(image, size: .zero, baseline: 0)
+  }
+}
